refactor(api): extract live routes to src/api/live_routes.py [Patch 9/9]
Move 7 live-trading routes, 10 Pydantic models, live-runner state, and
all live-exclusive helpers from api_server.py into a self-contained
route module following the PR #375 pattern.

- live_routes.py defines CommitMandateRequest, LiveHaltRequest,
  ActiveMandateState, MandateLimits, LiveRunnerUnavailable, etc. locally
- Module-level state: _runner_tasks, _runner_factory
- register_live_routes() resolves require_auth, _validate_path_param,
  _get_session_service via sys.modules fallback
- 7 routes: mandate/commit, live/halt, live/resume, live/status,
  live/authorize, live/runner/start, live/runner/stop
- 22 symbols re-exported for test monkeypatch compat

Also fix sessions_routes.py: move SSE frame helpers (mandate_proposal,
live_action) to module level so test_api_live_runtime.py can access them
via api_server re-exports.

api_server.py: 1811 -> 1102 lines (-709)

AI-Model: idealab/Qwen3.7-Max-DogFooding

AI-Contributed/Feature: 1463/1463
AI-Contributed/UT: 0/0
