+++ agent/api_server.py
@@ -142,126 +142,7 @@ class RunResponse(BaseModel):
 # Session/goal Pydantic models are defined in src/api/sessions_routes.py.
 
 
-# ---- Live trading channel: consent commit + kill switch ----
-
-
-class CommitMandateRequest(BaseModel):
-    """Surface-originated mandate commit (Consent §1 / §3).
-
-    This is the ONLY write path that activates a live-trading mandate. It is a
-    privileged HTTP action the user surface sends on an explicit click/keypress
-    — NOT a tool the agent model can call. ``consent_ack`` MUST be ``true``.
-    """
-
-    broker: str = Field(..., min_length=1, max_length=64)
-    proposal_id: str = Field(..., pattern=r"^mp_[0-9a-f]{32}$")
-    selected_ordinal: int = Field(..., ge=1, le=10)
-    adjustments: Optional[Dict[str, Any]] = None
-    consent_ack: bool = Field(..., description="Explicit affirmative; must be true")
-    session_id: Optional[str] = None
-    account_ref: str = Field("", max_length=128)
-    lifetime_days: int = Field(30, ge=1, le=365)
-
-
-class LiveHaltRequest(BaseModel):
-    """Trip or clear the live kill switch (Consent §4).
-
-    Tripping/clearing is a privileged surface action, never an agent tool. When
-    ``broker`` is omitted the GLOBAL switch is used (halts every broker).
-    """
-
-    broker: Optional[str] = Field(None, max_length=64)
-    reason: str = Field("user requested halt", max_length=500)
-    session_id: Optional[str] = None
-
-
-class LiveAuthorizeRequest(BaseModel):
-    """Kick off (or describe) the OAuth bootstrap for a live broker (C2).
-
-    Vibe-Trading never holds funds and never operates a venue, so the OAuth
-    bootstrap runs through the broker's own user-authorized device flow on the
-    client (CLI / desktop MCP), not a server-side redirect. This endpoint is the
-    web on-ramp: it tells a Web UI user exactly how to discover/start the flow.
-    """
-
-    broker: str = Field(..., min_length=1, max_length=64)
-
-
-class LiveRunnerControlRequest(BaseModel):
-    """Start or stop the persistent live runner for one broker (SPEC §7.5).
-
-    The runner wakes on schedule/market events and trades autonomously inside a
-    committed mandate. Starting it is a privileged surface action, never an
-    agent tool. A committed, unexpired mandate must already exist.
-    """
-
-    broker: str = Field(..., min_length=1, max_length=64)
-    session_id: Optional[str] = None
-
-
-class BrokerAuthState(BaseModel):
-    """Per-broker authorization snapshot for ``GET /live/status``."""
-
-    broker: str
-    oauth_token_present: bool = Field(..., description="Whether an OAuth token cache exists")
-    is_live_broker: bool = Field(..., description="Whether this key is a recognized live broker")
-
-
-class MandateLimits(BaseModel):
-    """Flattened active-mandate limits surfaced to the UI (Mandate layer a/b)."""
-
-    max_order_notional_usd: float
-    max_total_exposure_usd: float
-    max_leverage: float
-    max_trades_per_day: int
-    allowed_instruments: List[str]
-    account_funding_usd: float
-
-
-class ActiveMandateState(BaseModel):
-    """Active-mandate snapshot with the expiry countdown (SPEC §9 dec. 2)."""
-
-    broker: str
-    account_ref: str
-    created_at: str
-    expires_at: str
-    expires_in_seconds: Optional[int] = Field(
-        None, description="Seconds until expiry; negative when already expired"
-    )
-    expired: bool
-    limits: MandateLimits
-
-
-class RunnerLivenessState(BaseModel):
-    """Runner liveness snapshot via the §7.5 liveness contract."""
-
-    broker: str
-    alive: bool
-    last_tick: Optional[float] = Field(None, description="Unix epoch of last heartbeat tick")
-    last_tick_age_seconds: Optional[float] = None
-
-
-class LiveBrokerStatus(BaseModel):
-    """Combined live-channel status for a single broker."""
-
-    auth: BrokerAuthState
-    mandate: Optional[ActiveMandateState] = None
-    runner: RunnerLivenessState
-    halted: bool = Field(..., description="Per-broker OR global kill switch is tripped")
-
-
-class LiveStatusResponse(BaseModel):
-    """Top-level live-channel status (C2)."""
-
-    global_halted: bool = Field(..., description="Whether the GLOBAL kill switch is tripped")
-    brokers: List[LiveBrokerStatus]
-
-
-class ChannelPairingCommandRequest(BaseModel):
-    """Pairing command executed through the IM control surface."""
-
-    channel: str = Field(..., min_length=1, max_length=64)
-    command: str = Field("list", max_length=500)
+# Live-trading Pydantic models are defined in src/api/live_routes.py.
 
 
 # ============================================================================
@@ -1072,634 +953,38 @@ from src.api.swarm_routes import _get_swarm_runtime  # noqa: F401, E402
 
 
 # ============================================================================
-# Live trading channel — consent commit + kill switch
+# Live trading routes - defined in src/api/live_routes.py
 # ============================================================================
-#
-# These are the privileged SURFACE actions of the live-trading channel
-# (live-trading SPEC, Consent §1/§3/§4). None is an agent tool:
-#   - POST /mandate/commit  -> the single mandate writer (commit_mandate)
-#   - POST /live/halt       -> trip the kill switch (P5 trip_halt)
-#   - POST /live/resume     -> clear the kill switch (P5 clear_halt)
-# Each best-effort relays a mandate.committed / live.halted / live.action event
-# through the EXISTING session EventBus, so the frontend's already-wired
-# /sessions/{id}/events SSE stream reflects the state change. No new bus.
 
-
-def _emit_live_event(session_id: Optional[str], event_type: str, data: Dict[str, Any]) -> None:
-    """Best-effort relay of a live-channel event through the existing bus.
-
-    The event flows out the existing ``/sessions/{session_id}/events`` SSE
-    stream. Notifications never gate autonomy (SPEC Consent §5): a relay failure
-    or a missing session is swallowed — the state change already happened on disk.
-
-    Args:
-        session_id: Target session, or ``None`` to skip relay.
-        event_type: SSE event name (``mandate.committed`` / ``live.halted`` /
-            ``live.resumed`` / ``live.action``).
-        data: JSON-serializable event payload.
-    """
-    if not session_id:
-        return
-    try:
-        svc = _get_session_service()
-        if svc and svc.get_session(session_id):
-            svc.event_bus.emit(session_id, event_type, data)
-    except Exception:  # pragma: no cover - relay is non-blocking by contract
-        logger.debug("live event relay failed for %s/%s", session_id, event_type, exc_info=True)
-
-
-
-
-
-def _fetch_broker_ceilings(broker: str) -> Optional[Dict[str, Any]]:
-    """Best-effort fetch of broker-side account ceilings for the commit re-check.
-
-    Reads the broker's mapped account/portfolio tool and derives an authoritative
-    ceiling snapshot (buying power / funding) so the commit-time fit check binds
-    to the venue's real limits rather than an agent-proposed number. Returns
-    ``None`` on any failure (channel not configured, tool error, fields not
-    recognized) so the caller falls back to the proposal's own snapshot — a
-    commit is never blocked on a broker read.
-
-    Args:
-        broker: The live-broker key.
-
-    Returns:
-        A ceilings dict (canonical keys) or ``None`` to fall back.
-    """
-    try:
-        adapter = _live_broker_adapter(broker)
-    except LiveRunnerUnavailable:
-        return None
-    try:
-        from src.trading.service import runner_tool_name
-
-        account_tool = runner_tool_name(broker, "account") or "get_account"
-        result = adapter.call_tool(account_tool, {})
-    except Exception:  # pragma: no cover - status/commit must never raise here
-        logger.debug("broker ceiling fetch failed for %s", broker, exc_info=True)
-        return None
-    if not isinstance(result, dict) or result.get("status") == "error":
-        return None
-    payload = result.get("result") if isinstance(result.get("result"), dict) else result
-    funding: Optional[float] = None
-    for key in ("account_funding_usd", "buying_power", "cash", "portfolio_value", "equity"):
-        raw = payload.get(key) if isinstance(payload, dict) else None
-        try:
-            if raw is not None:
-                funding = float(raw)
-                break
-        except (TypeError, ValueError):
-            continue
-    if funding is None or funding <= 0:
-        return None
-    # A single order can never exceed available funding; total exposure is capped
-    # at funding for a cash account. Leverage stays at 1.0 unless the broker
-    # reports margin (L6). These canonical keys are normalized by commit_mandate.
-    return {
-        "account_funding_usd": funding,
-        "max_order_notional_usd": funding,
-        "max_total_exposure_usd": funding,
-    }
-
-
-@app.post("/mandate/commit", dependencies=[Depends(require_auth)])
-async def commit_mandate_endpoint(payload: CommitMandateRequest):
-    """Commit a user-selected mandate profile — the only mandate write path.
-
-    Calls :func:`src.live.mandate.commit.commit_mandate`, which re-validates the
-    proposal is live and the resolved profile still fits the ceilings the user
-    saw. Requires ``consent_ack=true`` (rejected otherwise). On success emits a
-    ``mandate.committed`` + ``live.action`` event so all surfaces reflect the
-    newly active mandate.
-    """
-    if payload.consent_ack is not True:
-        raise HTTPException(status_code=400, detail="consent_ack must be true to commit a mandate")
-
-    from src.live.mandate.commit import CommitError, commit_mandate
-
-    # Prefer broker-DERIVED ceilings over the agent-supplied proposal snapshot:
-    # the commit re-check should bind to the venue's real account limits, not a
-    # number the model proposed. Best-effort — falls back to the proposal's own
-    # ceilings (commit_mandate handles ceilings_ref=None) when the broker channel
-    # is unavailable or the read fails (we never block a commit on a broker read).
-    broker_ceilings = _fetch_broker_ceilings(payload.broker)
-
-    try:
-        result = commit_mandate(
-            proposal_id=payload.proposal_id,
-            ordinal=payload.selected_ordinal,
-            adjustments=payload.adjustments,
-            consent_ack=payload.consent_ack,
-            broker=payload.broker,
-            account_ref=payload.account_ref,
-            session_id=payload.session_id,
-            ceilings_ref=broker_ceilings,
-            lifetime_days=payload.lifetime_days,
-        )
-    except CommitError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    _emit_live_event(payload.session_id, "mandate.committed", result)
-    _emit_live_event(
-        payload.session_id,
-        "live.action",
-        {"kind": "mandate_committed", "broker": result["broker"], "mandate_id": result["mandate_id"]},
-    )
-    return result
-
-
-@app.post("/live/halt", dependencies=[Depends(require_auth)])
-async def halt_live_endpoint(payload: LiveHaltRequest):
-    """Trip the live kill switch (privileged surface action, Consent §4).
-
-    Writes the HALT sentinel via :func:`src.live.halt.trip_halt`; the
-    enforcement gate then rejects every order attempt until resumed. Emits a
-    ``live.halted`` event so all surfaces reflect the halted state.
-    """
-    from src.live.halt import trip_halt
-
-    try:
-        path = trip_halt(by="frontend", reason=payload.reason, broker=payload.broker)
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    result = {"halted": True, "broker": payload.broker, "reason": payload.reason, "sentinel": str(path)}
-    _emit_live_event(payload.session_id, "live.halted", result)
-    _emit_live_event(
-        payload.session_id,
-        "live.action",
-        {"kind": "halt_tripped", "broker": payload.broker, "reason": payload.reason},
-    )
-    return result
-
-
-@app.post("/live/resume", dependencies=[Depends(require_auth)])
-async def resume_live_endpoint(payload: LiveHaltRequest):
-    """Clear the live kill switch (privileged surface action, Consent §4).
-
-    Deletes the HALT sentinel via :func:`src.live.halt.clear_halt` (an explicit
-    re-enable; never an agent tool). Emits a ``live.resumed`` event.
-    """
-    from src.live.halt import clear_halt
-
-    try:
-        cleared = clear_halt(broker=payload.broker)
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    result = {"halted": False, "broker": payload.broker, "cleared": cleared}
-    _emit_live_event(payload.session_id, "live.resumed", result)
-    _emit_live_event(
-        payload.session_id,
-        "live.action",
-        {"kind": "halt_cleared", "broker": payload.broker, "cleared": cleared},
-    )
-    return result
-
-
-# ============================================================================
-# Live trading channel — status, authorize on-ramp, runner control (C2 + §7.5)
-# ============================================================================
-#
-# C2 surfaces the dormant-by-default channel state so a user can SEE what is and
-# is not authorized before trusting it: per-broker OAuth presence, the active
-# mandate with its expiry countdown, runner liveness, and the kill-switch state.
-# The runner-control endpoints start/stop the persistent §7.5 runner that trades
-# autonomously inside a committed mandate. None of these is an agent tool; they
-# are privileged surface actions like /mandate/commit and /live/halt.
-
-
-def _known_live_brokers() -> List[str]:
-    """Return the recognized live-broker keys (SPEC §7.2)."""
-    from src.config.schema import LIVE_BROKER_SERVER_KEYS
-
-    return sorted(LIVE_BROKER_SERVER_KEYS)
-
-
-def _oauth_token_present(broker: str) -> bool:
-    """Return whether an OAuth token cache exists for a broker (C2 auth state).
-
-    The token cache lives at ``<runtime_root>/live/<broker>/oauth/`` (0700) and
-    is created only when the user OAuth-authorizes the channel. A missing or
-    empty directory means the channel is dormant (read-only, no live path).
-    """
-    try:
-        from src.live.paths import broker_dir
-
-        oauth_dir = broker_dir(broker) / "oauth"
-        return oauth_dir.is_dir() and any(oauth_dir.iterdir())
-    except Exception:  # pragma: no cover - status must never raise
-        logger.debug("oauth presence check failed for %s", broker, exc_info=True)
-        return False
-
-
-def _active_mandate_state(broker: str) -> Optional[ActiveMandateState]:
-    """Build the active-mandate snapshot for a broker, or ``None`` when absent.
-
-    Reads the committed mandate via the frozen store contract and computes the
-    ``expires_at`` countdown (SPEC §9 dec. 2). A mandate whose ``expires_at`` has
-    passed is still surfaced, flagged ``expired`` so the UI can prompt re-consent.
-    """
-    from src.live.mandate.store import load_mandate
-
-    mandate = load_mandate(broker)
-    if mandate is None:
-        return None
-
-    consent = mandate.consent
-    caps = mandate.hard_caps
-    expires_in: Optional[int] = None
-    expired = False
-    try:
-        expires_dt = datetime.fromisoformat(consent.expires_at.replace("Z", "+00:00"))
-        from datetime import timezone
-
-        now = datetime.now(timezone.utc)
-        if expires_dt.tzinfo is None:
-            expires_dt = expires_dt.replace(tzinfo=timezone.utc)
-        delta = expires_dt - now
-        expires_in = int(delta.total_seconds())
-        expired = expires_in <= 0
-    except (ValueError, AttributeError):
-        logger.debug("could not parse expires_at for %s mandate", broker, exc_info=True)
-
-    return ActiveMandateState(
-        broker=broker,
-        account_ref=consent.account_ref,
-        created_at=consent.created_at,
-        expires_at=consent.expires_at,
-        expires_in_seconds=expires_in,
-        expired=expired,
-        limits=MandateLimits(
-            max_order_notional_usd=caps.max_order_notional_usd,
-            max_total_exposure_usd=caps.max_total_exposure_usd,
-            max_leverage=caps.max_leverage,
-            max_trades_per_day=caps.max_trades_per_day,
-            allowed_instruments=[str(getattr(i, "value", i)) for i in caps.allowed_instruments],
-            account_funding_usd=caps.account_funding_usd,
-        ),
-    )
-
-
-def _runner_liveness_state(broker: str) -> RunnerLivenessState:
-    """Build the runner-liveness snapshot for a broker (SPEC §7.5 contract).
-
-    Uses the §7.5 ``liveness`` module (``is_runner_alive`` / ``last_tick``),
-    keyed by broker as the runner id. The module is built concurrently (R1); a
-    missing module or any error is treated as "not alive" (fail-safe display).
-    """
-    alive = False
-    tick: Optional[float] = None
-    age: Optional[float] = None
-    try:
-        from src.live.runtime import liveness
-
-        alive = bool(liveness.is_runner_alive(broker))
-        raw_tick = liveness.last_tick(broker)
-        if raw_tick is not None:
-            tick = float(raw_tick)
-            age = max(0.0, time.time() - tick)
-    except Exception:  # pragma: no cover - liveness module is built concurrently
-        logger.debug("runner liveness lookup failed for %s", broker, exc_info=True)
-
-    return RunnerLivenessState(broker=broker, alive=alive, last_tick=tick, last_tick_age_seconds=age)
-
-
-@app.get("/live/status", response_model=LiveStatusResponse, dependencies=[Depends(require_auth)])
-async def live_status_endpoint(broker: Optional[str] = Query(None, max_length=64)):
-    """Return live-channel status: auth, active mandate, runner liveness, halt (C2).
-
-    Args:
-        broker: Optional single-broker filter. When omitted, every recognized
-            live broker is reported.
-
-    Returns:
-        A :class:`LiveStatusResponse` with the global kill-switch state and a
-        per-broker breakdown so the UI can show exactly what is authorized.
-    """
-    from src.live.halt import halt_flag_set
-
-    if broker is not None:
-        target = broker.strip().lower()
-        if not target:
-            raise HTTPException(status_code=400, detail="broker must not be blank")
-        brokers = [target]
-    else:
-        brokers = _known_live_brokers()
-
-    known = set(_known_live_brokers())
-    statuses: List[LiveBrokerStatus] = []
-    for key in brokers:
-        statuses.append(
-            LiveBrokerStatus(
-                auth=BrokerAuthState(
-                    broker=key,
-                    oauth_token_present=_oauth_token_present(key),
-                    is_live_broker=key in known,
-                ),
-                mandate=_active_mandate_state(key),
-                runner=_runner_liveness_state(key),
-                halted=halt_flag_set(broker=key),
-            )
-        )
-
-    return LiveStatusResponse(global_halted=halt_flag_set(broker=None), brokers=statuses)
-
-
-@app.post("/live/authorize", dependencies=[Depends(require_auth)])
-async def live_authorize_endpoint(payload: LiveAuthorizeRequest):
-    """Describe the OAuth bootstrap on-ramp for a live broker (C2 web on-ramp).
-
-    Vibe-Trading holds no funds and runs no venue: the OAuth flow happens on the
-    broker's own user-authorized device channel (CLI / desktop MCP), never a
-    server-side redirect. A Web UI user reaches this endpoint to DISCOVER how to
-    start the flow. It performs no authorization itself and never returns a token.
-    """
-    broker = payload.broker.strip().lower()
-    if not broker:
-        raise HTTPException(status_code=400, detail="broker must not be blank")
-    if broker not in set(_known_live_brokers()):
-        raise HTTPException(status_code=400, detail=f"unknown live broker: {broker}")
-
-    from src.trading.service import connector_profile_id_for_broker
-
-    connector_profile = connector_profile_id_for_broker(broker)
-    return {
-        "broker": broker,
-        "connector_profile": connector_profile,
-        "oauth_token_present": _oauth_token_present(broker),
-        "instruction": (
-            f"Run `vibe-trading connector authorize {connector_profile}` "
-            "from the device that will hold the broker session. This opens the "
-            "broker's own OAuth consent flow; Vibe-Trading never holds funds and "
-            "only relays intent once you authorize."
-        ),
-        "note": (
-            "The live channel stays read-only until the OAuth token is present AND a "
-            "mandate is committed AND order tools are explicitly enabled."
-        ),
-    }
-
-
-# ---- Runner control (SPEC §7.5): start / stop the persistent live runner ----
-#
-# A LiveRunner (R2 contract: ``LiveRunner(broker)`` with ``run_loop()`` /
-# ``run_once()``) is driven in a background task per broker. The factory is
-# injectable (``_runner_factory``) so tests stub it with no real agent/broker.
-# ``run_loop`` may be sync (long-blocking) or async; both are supported.
-
-_runner_tasks: Dict[str, "asyncio.Task[Any]"] = {}
-_runner_factory: Optional[Any] = None
-
-
-class LiveRunnerUnavailable(RuntimeError):
-    """Raised when a live runner cannot be wired (broker not configured/authorized).
-
-    Distinct from a programming error so the start endpoint can map it to a 503
-    rather than a 500: the runtime is fine, the broker channel just isn't ready.
-    """
-
-
-def _live_broker_adapter(broker: str) -> Any:
-    """Build an ``MCPServerAdapter`` for a live broker from the user-side config.
-
-    Resolves the broker's MCP server entry by config key OR by a live-broker URL
-    host (so an aliased key still resolves), mirroring the registry's detection.
-
-    Args:
-        broker: The live-broker key, e.g. ``"robinhood"``.
-
-    Returns:
-        A constructed :class:`MCPServerAdapter` for the broker's read/write tools.
-
-    Raises:
-        LiveRunnerUnavailable: When no MCP server is configured for the broker.
-    """
-    from src.config.loader import load_agent_config
-    from src.tools.mcp import MCPServerAdapter
-
-    try:
-        from src.config.schema import is_live_broker_entry
-    except Exception:  # pragma: no cover - older schema without URL detection
-        is_live_broker_entry = None  # type: ignore[assignment]
-
-    cfg = load_agent_config()
-    servers = getattr(cfg, "mcp_servers", {}) or {}
-    for name, server_cfg in servers.items():
-        is_match = name == broker
-        if not is_match and is_live_broker_entry is not None and broker == "robinhood":
-            try:
-                is_match = is_live_broker_entry(name, server_cfg)
-            except Exception:  # pragma: no cover
-                is_match = False
-        if is_match:
-            return MCPServerAdapter(name, server_cfg)
-    raise LiveRunnerUnavailable(f"no MCP server configured for live broker {broker!r}")
-
-
-def _build_live_runner(broker: str) -> Any:
-    """Construct a fully-wired ``LiveRunner`` for a broker (SPEC §7.5 R-INT).
-
-    Wires the runner to the real surfaces — the public ``SessionService`` agent
-    caller (never the protected loop internals), the broker's READ/WRITE MCP
-    tools, the R4 reconciler, the R1 scheduler, and R3 market-hours triggers —
-    and injects an audit ``event_callback`` so every autonomous live action is
-    broadcast as a ``live.action`` SSE event on the runner's session bus.
-
-    Args:
-        broker: The live-broker key.
-
-    Returns:
-        A runner object exposing ``run_loop`` / ``run_once`` (R2 contract).
-
-    Raises:
-        LiveRunnerUnavailable: When the broker channel is not configured.
-    """
-    if _runner_factory is not None:
-        return _runner_factory(broker)
-
-    from src.live.audit import write_live_action
-    from src.live.runtime.reconcile import reconcile
-    from src.live.runtime.runner import LiveRunner
-    from src.live.runtime.scheduler import Scheduler
-    from src.live.runtime.triggers import Trigger
-    from src.trading.service import runner_tool_name
-
-    def _tool(operation: str) -> str:
-        remote_tool = runner_tool_name(broker, operation)
-        if remote_tool is None:
-            raise LiveRunnerUnavailable(
-                f"live runner for {broker!r} does not define remote tool {operation!r}"
-            )
-        return remote_tool
-
-    positions_tool = _tool("positions")
-    balance_tool = _tool("account")
-    open_orders_tool = _tool("orders")
-    submit_order_tool = _tool("submit_order")
-    cancel_order_tool = _tool("cancel_order")
-    adapter = _live_broker_adapter(broker)  # raises LiveRunnerUnavailable if absent
-
-    def _read(remote_tool: str):
-        """A zero-arg broker READ callable bound to one remote tool."""
-        return lambda: adapter.call_tool(remote_tool, {})
-
-    def _submit(order: Dict[str, Any]) -> Dict[str, Any]:
-        # Route the flatten sweep's normalized order to the broker's write tools.
-        # Field mapping against the real Robinhood schema is finalized post-access
-        # (L6); the action discriminator is broker-agnostic.
-        if order.get("action") == "cancel":
-            return adapter.call_tool(cancel_order_tool, order)
-        return adapter.call_tool(submit_order_tool, order)
-
-    svc = _get_session_service()
-    session = svc.create_session(title=f"live-runner:{broker}")
-    session_id = session.session_id
-
-    async def _agent_caller(sid: str, prompt: str) -> Dict[str, Any]:
-        # Dispatch one autonomous turn through the PUBLIC SessionService entry.
-        # The agent then trades within the mandate via the gated order tools.
-        return await svc.send_message(sid, prompt)
-
-    def _audit_with_bus(event: Any) -> Dict[str, Any]:
-        # Broadcast each live action as a live.action SSE event on the runner's
-        # session bus (no protected-loop touch — the runner owns its session).
-        return write_live_action(
-            event,
-            event_callback=lambda etype, record: svc.event_bus.emit(session_id, etype, record),
-        )
-
-    # Wire the scheduler's fire callback to the runner's tick. The scheduler is
-    # constructed before the runner (it needs on_fire), and the runner needs the
-    # scheduler, so late-bind via a holder to break the cycle.
-    runner_holder: Dict[str, Any] = {}
-
-    async def _on_fire(_job: Any) -> None:
-        runner = runner_holder.get("runner")
-        if runner is not None:
-            await runner.run_once()
-
-    scheduler = Scheduler(_on_fire)
-
-    runner = LiveRunner(
-        broker,
-        agent_caller=_agent_caller,
-        reconcile_fn=reconcile,
-        read_positions=_read(positions_tool),
-        read_balance=_read(balance_tool),
-        read_open_orders=_read(open_orders_tool),
-        submit_fn=_submit,
-        write_audit_fn=_audit_with_bus,
-        scheduler=scheduler,
-        triggers=[Trigger.market("us_equity")],
-        session_id=session_id,
-    )
-    runner_holder["runner"] = runner
-    return runner
-
-
-async def _drive_runner(runner: Any) -> None:
-    """Run a runner's ``run_loop`` to completion, sync or async.
-
-    A synchronous ``run_loop`` is offloaded to a worker thread so it does not
-    block the event loop; an async ``run_loop`` is awaited directly.
-    """
-    result = runner.run_loop()
-    if asyncio.iscoroutine(result):
-        await result
-    else:
-        await asyncio.get_running_loop().run_in_executor(None, lambda: result)
-
-
-@app.post("/live/runner/start", dependencies=[Depends(require_auth)])
-async def start_runner_endpoint(payload: LiveRunnerControlRequest):
-    """Start the persistent live runner for a broker (SPEC §7.5).
-
-    Refuses to start unless a committed, unexpired mandate exists and the kill
-    switch is clear — the runner trades autonomously, so it must not start into a
-    dead/halted channel. Idempotent: a request for an already-running broker
-    returns ``already_running`` without spawning a second task.
-    """
-    from src.live.halt import halt_flag_set
-
-    broker = payload.broker.strip().lower()
-    if not broker:
-        raise HTTPException(status_code=400, detail="broker must not be blank")
-    from src.trading.service import broker_supports_live_runner
-
-    if not broker_supports_live_runner(broker):
-        raise HTTPException(
-            status_code=400,
-            detail=f"live runner is not supported for {broker}",
-        )
-
-    existing = _runner_tasks.get(broker)
-    if existing is not None and not existing.done():
-        return {"broker": broker, "started": False, "already_running": True}
-
-    mandate = _active_mandate_state(broker)
-    if mandate is None:
-        raise HTTPException(status_code=409, detail=f"no committed mandate for {broker}")
-    if mandate.expired:
-        raise HTTPException(status_code=409, detail=f"mandate for {broker} has expired; re-authorize first")
-    if halt_flag_set(broker=broker) or halt_flag_set(broker=None):
-        raise HTTPException(status_code=409, detail="kill switch is tripped; resume before starting the runner")
-
-    try:
-        runner = _build_live_runner(broker)
-    except LiveRunnerUnavailable as exc:
-        raise HTTPException(status_code=503, detail=str(exc)) from exc
-    except Exception as exc:
-        raise HTTPException(status_code=500, detail=f"could not construct runner: {exc}") from exc
-
-    task = asyncio.ensure_future(_drive_runner(runner))
-    _runner_tasks[broker] = task
-    task.add_done_callback(
-        lambda t, b=broker: _runner_tasks.pop(b, None) if _runner_tasks.get(b) is t else None
-    )
-
-    _emit_live_event(
-        payload.session_id,
-        "live.action",
-        {"kind": "runner_started", "broker": broker},
-    )
-    return {"broker": broker, "started": True, "already_running": False}
-
-
-@app.post("/live/runner/stop", dependencies=[Depends(require_auth)])
-async def stop_runner_endpoint(payload: LiveRunnerControlRequest):
-    """Stop the persistent live runner for a broker (SPEC §7.5).
-
-    Cancels the background task. This does NOT flatten positions — that is the
-    preemptive kill switch's job (``/live/halt`` -> flatten); stopping the runner
-    simply ceases new autonomous turns. Idempotent for an already-stopped broker.
-    """
-    broker = payload.broker.strip().lower()
-    if not broker:
-        raise HTTPException(status_code=400, detail="broker must not be blank")
-    from src.trading.service import broker_supports_live_runner
-
-    if not broker_supports_live_runner(broker):
-        raise HTTPException(
-            status_code=400,
-            detail=f"live runner is not supported for {broker}",
-        )
-
-    task = _runner_tasks.pop(broker, None)
-    if task is None or task.done():
-        return {"broker": broker, "stopped": False, "was_running": False}
-
-    task.cancel()
-    _emit_live_event(
-        payload.session_id,
-        "live.action",
-        {"kind": "runner_stopped", "broker": broker},
-    )
-    return {"broker": broker, "stopped": True, "was_running": True}
-
+from src.api.live_routes import register_live_routes  # noqa: E402
+
+register_live_routes(app)
+
+# Re-export for test monkeypatch compatibility
+from src.api.live_routes import (  # noqa: F401, E402
+    CommitMandateRequest,
+    LiveHaltRequest,
+    LiveAuthorizeRequest,
+    LiveRunnerControlRequest,
+    BrokerAuthState,
+    MandateLimits,
+    ActiveMandateState,
+    RunnerLivenessState,
+    LiveBrokerStatus,
+    LiveStatusResponse,
+    LiveRunnerUnavailable,
+    _runner_tasks,
+    _runner_factory,
+    _emit_live_event,
+    _fetch_broker_ceilings,
+    _known_live_brokers,
+    _oauth_token_present,
+    _active_mandate_state,
+    _runner_liveness_state,
+    _live_broker_adapter,
+    _build_live_runner,
+    _drive_runner,
+)
 
 # ============================================================================
 # Alpha Zoo routes (Web UI) — defined in src/api/alpha_routes.py
